Split library
