+++ faebryk/library/__init__.py
@@ -1,4 +1,3 @@
 import faebryk.library.core
-import faebryk.library.kicad
 import faebryk.library.library
 import faebryk.library.traits
