+++ faebryk/library/library/__init__.py
@@ -1,0 +1,9 @@
+# This file is part of the faebryk project
+# SPDX-License-Identifier: MIT
+
+import faebryk.library.library.components
+import faebryk.library.library.footprints
+import faebryk.library.library.interfaces
+import faebryk.library.library.kicad
+import faebryk.library.library.links
+import faebryk.library.library.parameters
