+++ faebryk/library/library/links.py
@@ -1,0 +1,2 @@
+from faebryk.library.core import Link
+from faebryk.library.traits import *
