+++ faebryk/library/traits/__init__.py
@@ -1,0 +1,8 @@
+# This file is part of the faebryk project
+# SPDX-License-Identifier: MIT
+
+import faebryk.library.traits.component
+import faebryk.library.traits.footprint
+import faebryk.library.traits.interface
+import faebryk.library.traits.link
+import faebryk.library.traits.parameter
