+++ faebryk/library/traits/link.py
@@ -1,0 +1,7 @@
+# This file is part of the faebryk project
+# SPDX-License-Identifier: MIT
+
+import logging
+logger = logging.getLogger("library")
+
+from faebryk.library.core import LinkTrait
